backend/requirements.txt
asgiref==3.5.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.16
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.3
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
oauthlib==3.2.1
pep8==1.7.1
pep8-naming==0.12.1
Pillow==9.0.1
pyartnet==0.8.3
pyception==0.1.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.5.0
python-docx==0.8.11
python-dotenv==0.20.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
testfixtures==6.18.3
tomli==2.0.0
typed-ast==1.5.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.9
webcolors==1.12
zipp==3.7.0
